fix(chi): use assign clone returned by http.Request.WithContext to address of the original http.Request
This commit changes Handler.handle's behavior, so that when the hub is added to the request's context, the original request (i.e. its pointer) is updated, instead of using a new variable which is given to ServeHTTP.
This has the major advantage, that middlewares called before chisentry will have access to the hub in the context, if they called ServerHTTP, as we updated the *http.Request's underlying value.